06-lab-complete/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
python-dotenv==1.0.1
openai==1.12.0
google-genai
